+++ src/enums/serviceAliases.ts
@@ -73,4 +73,5 @@ export default {
   [services.transitGatewayAttachment]: 'transitGatewayAttachments',
   [services.vpcEndpoint]: 'vpcEndpoints',
   [services.vpnConnection]: 'vpnConnections',
+  [services.vpcPeeringConnection]: 'vpcPeeringConnections',
 }