+++ src/properties/logger.ts
@@ -681,4 +681,8 @@ export default {
    */
    fetchedManagedPrefixLists: (num: number): string => `Found ${num} Managed Prefix Lists`,
    fetchedManagedPrefixListEntries: (num: number): string => `Found ${num} Managed Prefix List Entries`,
+  /**
+   * Vpc Peering Connections
+   */
+  fetchedVpcPeeringConnections: (num: number): string => `Found ${num} Vpc Peering Connections`,
 }