+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [0.82.0-alpha.3](https://github.com/cloudgraphdev/cloudgraph-provider-aws/compare/0.82.0-alpha.2...0.82.0-alpha.3) (2022-05-24)
+
+
+### Features
+
+* **services:** Add service vpcPeeringConnection ([b5bda76](https://github.com/cloudgraphdev/cloudgraph-provider-aws/commit/b5bda76de47268f12390b4f554067e6347187da3))
+
+# [0.82.0-alpha.2](https://github.com/cloudgraphdev/cloudgraph-provider-aws/compare/0.82.0-alpha.1...0.82.0-alpha.2) (2022-05-23)
+
+
+### Bug Fixes
+
+* **services:** fix nat gateway service data ([1d0303e](https://github.com/cloudgraphdev/cloudgraph-provider-aws/commit/1d0303ef27ec55a91eb27e015d877fbc428c211d))
+
+# [0.82.0-alpha.1](https://github.com/cloudgraphdev/cloudgraph-provider-aws/compare/0.81.1-alpha.1...0.82.0-alpha.1) (2022-05-23)
+
+
+### Features
+
+* Add vpcEndpoint service ([a33a61f](https://github.com/cloudgraphdev/cloudgraph-provider-aws/commit/a33a61f1b5cd9f0a4bc49682f0ef44e4d2880283))
+
+## [0.81.1-alpha.1](https://github.com/cloudgraphdev/cloudgraph-provider-aws/compare/0.81.0...0.81.1-alpha.1) (2022-05-23)
+
+
+### Bug Fixes
+
+* **services:** fix route table service mainRouteTable property ([e99287b](https://github.com/cloudgraphdev/cloudgraph-provider-aws/commit/e99287b91828e57715d240f23a5cc0276799da12))
+
 # [0.81.0](https://github.com/cloudgraphdev/cloudgraph-provider-aws/compare/0.80.0...0.81.0) (2022-05-19)